exns v0.3.5-beta Exns.Request.Worker

An Exns.Request.Worker is implemented as a GenServer, which is to be spawned by poolboy.

Its basic function is to open a socket to the remove service (on init) then make requests via this socket. The socket is not closed after each request, but rather on the GenServer’s termination.

Functions

new_socket(address)

new_socket(address, old_socket)

send_recv(state, encoded, ref, retries)

start_link(args, opts \\ [])